2022/07/07　変更

+++ evaluation.py
@@ -42,7 +42,7 @@ def draw_matches_cv(data, matches, plot_points=True):
     img2 = to3dim(data['image2'])
     img1 = np.concatenate([img1, img1, img1], axis=2)
     img2 = np.concatenate([img2, img2, img2], axis=2)
-    return cv2.drawMatches(img1, keypoints1, img2, keypoints2, matches,
+    return cv2.drawMatches(np.uint8(img1), keypoints1, np.uint8(img2), keypoints2, matches,
                            None, matchColor=(0,255,0), singlePointColor=(0, 0, 255))
 
 def isfloat(value):
@@ -367,16 +367,12 @@ def evaluate(args, **options):
                 draw_matches(image, warped_image, matches_temp, lw=1.0, 
                         filename=filename, show=False, if_fig=False)
 
-
-
-
-
-
     if args.repeatibility:
         repeatability_ave = np.array(repeatability).mean()
         localization_err_m = np.array(localization_err).mean()
         print("repeatability: ", repeatability_ave)
         print("localization error over ", len(localization_err), " images : ", localization_err_m)
+    
     if args.homography:
         correctness_ave = np.array(correctness).mean(axis=0)
         # est_H_mean_dist = np.array(est_H_mean_dist)
@@ -390,8 +386,6 @@ def evaluate(args, **options):
             print("mean AP", mAP_m)
 
         print("end")
-
-
 
     # save to files
     with open(save_file, "a") as myfile:
